Handle Github branch protection rule webhook events
Github now emits a webhook event when a branch protection rule changes.

Add support for that in the Github driver.  We will update the branch
protection status in the branch cache immediately, and emit a trigger
event which will cause tenant reconfigurations if the protection status
of a branch has changed in any tenants.

Note: a branch protection rule applies to any number of branches, so
we may generate multiple Zuul trigger events from a single Github
webhook event in this case.

Change-Id: I0a7af786f9c69cf67eaaf4c75f437f8cf64a054a