Inject bundle id to inventroy zuul.items
Current usecase:
We want to skip job runs if the change is in a bundle (circular depency) and
it is not the first change in the same project of the bundle.
The solution is to create a parent job (dependent) what get the list of zuu
items from zuul vars and need the bunld id in the items to check that.

Change-Id: Ie437f79d7d2e7d9a372f7d81505d851cde1fd425