Update patch set 3
Patch Set 3:

One of the long-held concerns I have (and others seem to share) is that it's very hard to reason about irrelevant-files behavior. As you've observed, combining files and irrelevant-files in the same job makes no sense, and worse you can easily end up inheriting one from a parent and so accidentally combine it with the other without it being immediately obvious. You'll notice that Zuul's own job definitions rely only on the files attribute and avoid use of irrelevant-files entirely. Unfortunately, it doesn't look like our current documentation makes that risk particularly clear.

What I like about the proposed implementation is that it avoids the pitfall of users accidentally combining files and irrelevant-files, since the entire fileset is overridden if also set in a child job. I also appreciate that an empty includes list is treated as including everything rather than including nothing; our philosophy with Zuul's job selection behavior is to prefer running a job over not running it when there's any doubt as to what the user might have meant. That also makes configuration more terse if the user only wants to exclude files, though I don't see any examples where you're explicitly testing that behavior (apologies if I overlooked it).

My primary concern with this is that it, similar to irrelevant-files, offers an attractive option with hidden complexities which may allow users to get themselves into trouble and end up not running jobs when they meant to. I agree though that it's at least got a clear set of rules users can reason through, as long as those are documented.

One thing I would like to see is documentation for the new job attribute in doc/source/config/job.html, perhaps including an explanation of how to replicate the files and irrelevant-files behaviors in filesets since this could provide a path for migration if we were to decide to deprecate them in favor of this solution. An up-side to including documentation within the change is that reviewers can more easily understand the feature without needing to thoroughly explain the same points in the commit message (which was very detailed though, thank you).

Patch-set: 3
CC: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>