Update patch set 3
Patch Set 3:

As discussed, files and irrelevant-files are rather complex, and I don't think we should add to the confusion by supporting both.  If we are to do so, then I think we should deprecate and remove files/irrelevant-files.  It would be good to have a plan for this, a timeframe, and documentation and/or tools to help people with the conversion.

Getting buy-in on the mailing list or via a spec would be good.

Patch-set: 3
Attention: {"person_ident":"Gerrit User 1 \u003c1@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"REMOVE","reason":"removed on reply"}