Use JobData for Build result data
The result data from a job may be quite large, which is why we
offload it to a sharded ResultData object.  It may also be
nonexistent, which is why we just set it to null and skip the
offloading in that case.

The JobData class that we use for FrozenJobs has two extra
optimizations: if the data size is small, we just store it
inline instead of creating a small single-node shard; and if
we do shard, we store a hash inline so we only refresh if
necessary.

Update the Build object to use JobData instances instead of
ResultData for the build result data so that we can benefit
from the additional optimizations.

This is, surprisingly, the first time we have had to check the
cluster model api within model.py, so this change also performs
a slight rearrangement of the MODEL_API constant to avoid
circular imports.

Change-Id: I2053bd0327dfcad1d4246b158098033e188e2ed4