Trigger pipeline runs when semaphores release
The recent optimization to avoid processing pipelines if no events are
waiting did not account for semaphores which may be held by jobs in
different pipelines.  In that case, a job completing in one pipeline
needs to generate an event in another pipeline in order to prompt it
to begin processing.

We have no easy way of knowing which pipelines may have jobs which are
waiting for a semaphore, so this change broadcasts an event to every
pipeline in the tenant when a semaphore is released.  Hopefully this
shouldn't generate that much more traffic (how much depends on how
frequently semaphores are released).  If desired, we can further
optimize this by storing semaphore pipeline waiters in ZK in a later
change.

Change-Id: Ide381279b0442d11535c00746e4baf19f32f3cd7