Support submitWholeTopic in Gerrit
This adds a query for changes which are set (by Gerrit) to be submitted
together due to submitWholeTopic being enabled.  In this case, changes
which are of the same topic will be merged simultaneously by Gerrit,
therefore Zuul should treat them as a set of circular dependencies.

The behavior is automatic, since Gerrit will return a set of changes
if the setting is enabled, or the empty list if it is not.  Zuul still
requires that circular dependencies be enabled in any queues where they
appear.

If users have submitWholeTopic enabled in Gerrit but do not have
allow-circular-dependencies enabled, they may begin to see errors.
However, the errors are self-explanatory, and installations such as
these are already not testing what Gerrit will merge, so reporting
the discrepancy is an improvement.

Change-Id: Icf65913a049a9b9ddbedd20cc73bf44ffcc831b8