web/api: Reject unknown methods and consolidate CORS and request checking
This change is a combination of four older changes [1,2,3,4] which
received comments which received comments that have been squished into
this single change.

This was inspired when looking at haproxy liveness checks, where it
would make OPTIONS requests that Zuul was returning as complete
responses.

Since we already have a wrapper checking the OPTIONS request type, it
seems to make sense to consolidate all method checking into this one
location.  This wrapper is renamed to
"handle_cors_preflight_and_filter_methods" as ... that is what it
does.

Exported functions are updated to provide the methods they support via
this wrapper.  In a couple of places in the router we change the match
so that we can filter out invalid request types.

This also makes more consistent the acceptance of the HEAD method,
which is now explicitly listed.  Additionally, consenus from reading a
few things seems to be that you don't need to specify OPTIONS in
Access-Control-Allow-Methods because the only way to get that header
back is via the OPTIONS request, so it is implied.

A couple of places in the tests assume 401 for things that are now 405
(i.e. this will never work, instead of unauthorized); these are
updated.

[1] I736cef084a3239f4d0a5a3b64c21e42c11631d5d
[2] I6ee2e4807e158884f17809f366ce9b59397c95e5
[3] I2a38ecadac6eb56c8c485b39582ab99ce60cbed8
[4] Ib292219e598f6923d2abff4950fe0730a5bcbcb7

Change-Id: I3f46344ad7a2688f93d9395377b6fe5d0ab9b906