zuul/web : Make /project end-point reject PUT
When no authenication is setup, the authenticated end-points like

 api/tenant/tenant-one/project/org/project/<...>

fall back to the /project handler.  This handler does not accept POST
requests, and should return a 504 for any !GET requests.  This update
is made.

This requires a slight adjustment in the test-case; currently it is
checking the authenticated end-points return a 404 when
unauthenticated.  As noted, these requests fall back into /project, so
PUT requests will return a 504 (as they can't be handled) and GET
requests will return a 404 (as there's nothing there).  The test is
updated to try both methods.

Change-Id: I6ee2e4807e158884f17809f366ce9b59397c95e5