Include original path of renamed file for a PR
When a file is moved/renamed Github will only return an entry for the
file with the new name. However, the previous path also needs to be
included in the list of files. This is especially important when a Zuul
config file is renamed but also when `job.[irrelevant-]files` is used.

Change-Id: Ieba250bed57c8a9c2e7811476c202d530f2b30f1