zuul/web : move method checking into common decorator
This is a follow-on to I736cef084a3239f4d0a5a3b64c21e42c11631d5d that
moves the valid-method checking for exposed functions into a common
decorator that now responds to CORS pre-flight requests and also
filters correct methods for the end-point.

The decorator is added to all functions, and any unnecessary
in-function method checks are removed.  I think that existing testing
is sufficient, as we already have tests that hit these end-points with
invalid methods.

Change-Id: Ib292219e598f6923d2abff4950fe0730a5bcbcb7