Fix change list corruption on local layout update
When a pipeline's change list is updated while a post-config hook is
executed we can end up with a corrupted change list.

Since the pipeline change list is a sharded ZKObject it must be deleted
before the new content is written. If this races with the change list's
`create()` call we end up with shards from different writes.

To fix this we only create the `PipelineChangeList` object without
trying to create Znode when we don't hold the pipeline lock.

2022-03-11 14:16:21,658 ERROR zuul.Scheduler: Exception processing pipeline check in tenant foobar
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/zkobject.py", line 350, in _load
    self._set(**self.deserialize(data, context))
  File "/opt/zuul/lib/python3.8/site-packages/zuul/model.py", line 849, in deserialize
    data = super().deserialize(data, context)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/zkobject.py", line 87, in deserialize
    return json.loads(data.decode('utf-8'))
  File "/usr/local/lib/python3.8/json/__init__.py", line 357, in loads
    return _default_decoder.decode(s)
  File "/usr/local/lib/python3.8/json/decoder.py", line 340, in decode
    raise JSONDecodeError("Extra data", s, end)
json.decoder.JSONDecodeError: Extra data: line 1 column 592 (char 591)
The above exception was the direct cause of the following exception:
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 1933, in process_pipelines
    refreshed = self._process_pipeline(tenant, pipeline)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/scheduler.py", line 1972, in _process_pipeline
    pipeline.change_list.refresh(ctx)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/model.py", line 810, in refresh
    super().refresh(context)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/zkobject.py", line 156, in refresh
    self._load(context)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/zkobject.py", line 375, in _load
    raise InvalidObjectError from exc
zuul.zk.exceptions.InvalidObjectError

Change-Id: I848d5bb3616967971aa25b79d9905a43c95560d7