Clear branch cache on full reconfiguration
The full reconfiguration event is primarily used to correct invalid
states which may result from missed events, such as a branch creation.
However, it did not invalidate or update the actual data in the branch
cache.  Correct this by completely clearing the branch cache of every
connection when starting a full reconfiguration.  This will cause the
connections to query their upstream sources the next time that we ask
for a list of branches.

Change-Id: I1529fe1dd6092de612a94c14ec0527ab76f62e47