Update patch set 3
Patch Set 3: Code-Review+2

At a high level, I'm wholly in favor of this. The approach we took early in 2.5+/3.x to fork Ansible and neuter many of its modules is error-prone at best. It may have added a little security for the first version of Ansible we forked (I'd debate that), but like all blacklist approaches it's hubris to assume you won't miss anything a potential abuser might spot. Forward-porting it without a complete re-auditing of the Ansible module set has led to a number of cases where we missed some new functionality in a module we were previously allowing, or failed to spot the addition of new modules with risky capabilities; and all it takes is one hole there to render the entire thing a waste of effort.

Approaching security from a cost/benefit perspective, the cost of this implementation both to our maintainers (in supporting it) and our users (inconvenience and inability to truly reuse their production playbooks in testing) seems to vastly outweigh what little security we might be getting from it, if any. It was an interesting experiment, but we have plenty of evidence at this point that it's not a solution.

I do think better securing Zuul is important, and that we should start from information security fundamentals like the principle of least privilege to look for new architectural improvements, compartmentalizations, et cetera which could further improve the security posture. I don't think removing our Ansible forks is appreciably reducing our security posture, however, merely getting rid of a false sense of security which we may have been imparting to our users unfairly.

Patch-set: 3
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2