Rely on the unparsed config cache in reconfigurations
When performing a tenant reconfiguration, we check the zk-based
file cache for every project-branch to see if we need to refresh
our unparsed config cache from the contents in ZK.  However, we
know which project-branches have changed.  We can assume that if
the unparsed branch cache is valid for for a project-branch ltime
(it won't be for projects that trigger tenant reconfig), then we
can use its contents without checking the ltime of the file cache
in ZK.  This will typically save a few minutes worth of ZK traffic
during tenant reconfigurations on large tenants.

In order to make the background layout update similarly quick, we
need to prove the same min_ltimes to that process which were used
by the last reconfig.  To do this, we store the min_ltimes in ZK
for each reconfig (as sharded data).

Change-Id: Ia0a524053a110e8e48f709d219651e2ad9c8513d