Add a blob store and store large secrets in it
To optimize for the case where the same large secrets are used in
every job, this adds a global blob store and offloads secrets into
it.

Because this adds extra ZK operations, it is only used for secrets
which are > 10kb.

The blob store is pruned every hour by looking for entries which
are not in use and also have not been referenced since the start
of the cleanup process.

Change-Id: I43f08bc525d7f7b98ba74727c534489339432e8e