Fix recursive Gerrit change query
We have a history memo list which we use to avoid infinite recursion
in the Gerrit change query, however instead of mutating it, we copy
it.  This avoids infinite recursion in some circumstances, but at the
cost of exponential repitition.  It may also allow infinite recursion
depending on the order of results in Gerrit query responses.

To correct this, mutate the memo list instead of duplicating it.

Change-Id: Ic892cf5aa2db2f5e11f8562f857d9ec0a6b60c0a