De-duplicate change dependencies
Gerrit will return git parents in a submitted together response,
regardless of whether submitWholeTopic is set.  Make sure we emulate
that in our tests and validate the outcome.

Adding these tests has shown that Zuul is unable to enqueue dependency
cycles when dependencies are duplicated, so deduplicate them in the
model class.  This is done as late as possible so that dependencies
from as many sources as are available are de-duplicated.

Change-Id: Iaf94fc8947975ecbb615dbbc1c20a885835f69cd