Fix module loading for Windows jobs
The recent security fix that caused loading ansible.bultin.command
to load the normal command module was a little too aggressive since
the same mechanism is used by Ansible to load Windows powershell
support code, which is under the Ansible.ModuleUtils.* hierarchy.

This would result in an error from Ansible when attempting to run
the setup playbook:

  Could not find imported module support code for 'Ansible.ModuleUtils.Legacy'"'

This is corrected by reducing the scope of the mutation to just
ansible.builtin and ansible.legacy.

Change-Id: I70e9481478a3326692cb848ce0782f5331dc4758