Fix implied branch matchers with regex chars
If a project had a branch with a "+" in the name (or some other
special regex chars), then implied branch matchers would not match
the branch and jobs would not run.  Instead of treating implied
branch matchers as regexes, treat them as requiring exact matches.

To facilitate creating the correct type of branch matcher, create
them in the configloader instead of in the model classes.

Change-Id: I0421c5a446d0b75194096b1d8f0e0866dae3b8f0