Handle greedy project name regexes
Some users (and our documentation [0]) may want to have project name
regexes that match all projects: ^.*$. If the zuul installation has
colliding project names this is always an error because we evaluate the
non canonical names first then check for collisions.

To make these regexes work we evaluate canonical project name matches
first. If we find any matches with canonical names we do not check the
non canonical names. If there are no canonical name matches we check the
short names for matches and if found check for collisions.

[0] https://zuul-ci.org/docs/zuul/latest/tutorials/quick-start.html#configure-zuul-pipelines

Change-Id: I48fbab74fb42ba7c8005ef1a07b61c090acc154d