Handle greedy project name regexes
Some users (and our documentation [0]) may want to have project name
regexes that match all projects: ^.*$. If the zuul installation has
colliding project names this is always an error because we evaluate the
non canonical names first then check for collisions.

We have realized the point of the regex system is to simplify
configuration and apply configs to any projects that match. Collisions
don't impact this. Stop returning an error if there are collisions and
instead apply the configs to those projects.

[0] https://zuul-ci.org/docs/zuul/latest/tutorials/quick-start.html#configure-zuul-pipelines

Change-Id: I48fbab74fb42ba7c8005ef1a07b61c090acc154d