Skip reconfig when all project config is excluded
We've noticed that an update to a project that excludes all config will
still trigger a tenant reconfiguration. As this won't have any effect on
the actual tenant configuration we don't submit a reconfiguration event
if the project excludes all config classes.

Since every affected tenant will independently decide if a reconfig is
necessary, this won't cause problems for when a project is part of
multiple tenants.

Change-Id: Ida1a3167bc41397a93db2b2ae876f5e9588caf45