Only try to revise unaccepted node requests
When a node request was locked and accepted by a provider we can no
longer revise it. Since the schedulers try to update all but the
fulfilled requests, we can see a lot of "Unable to revise locked node
request ..." messages in our logs.

With this change we will only attempt to revise a request when it's
still in state 'requested'.

Change-Id: I4760ab3e4337026d60f840bec677b9fd9a67a573