Update change deps on refresh after cache miss
If a change key can not be found in the change cache and needs to be
refreshed from the source we also need to update the commit
dependencies.

Traceback (most recent call last):
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/scheduler.py", line 1988, in process_pipelines
    refreshed = self._process_pipeline(
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/scheduler.py", line 2056, in _process_pipeline
    self.process_pipeline_trigger_queue(tenant, pipeline)
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/scheduler.py", line 2218, in process_pipeline_trigger_queue
    self._process_trigger_event(tenant, pipeline, event)
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/scheduler.py", line 2248, in _process_trigger_event
    pipeline.manager.refreshDeps(change, event)
  File "/home/westphahl/src/opendev/zuul/zuul/zuul/manager/__init__.py", line 291, in refreshDeps
    for dep_change_ref in item.change.commit_needs_changes:
TypeError: 'NoneType' object is not iterable

Since resolving the dependencies requires a fully initialized tenant
with a layout we are calling the `_postConfig()` now after the layout is
assigned to the tenant.

Change-Id: I1c102b338ae6de05f3a752eb79f65744869e1860