client: generate a token silently if possible and needed
As a facility for local zuul admins, and in order to emulate the
behavior that was observed before the gearman cut, let the admin
CLI generate an auth token automatically from the configuration,
when calling a privileged subcommand without the --auth-token
argument.

The first authenticator found in the config is used.

Change-Id: Id8b0e45c7ca97a9f661ea6e71e378d1a67c8c454