Update patch set 1
Patch Set 1:

This patch series is structured to facilitate review.  There's a lot of boilerplate file operations in later patches, each of which should be contained to a patch which does only that thing, so you can review that without having to hunt for a changed line among lots of deleted lines.

As a consequence, the tests and documentation are generally removed before the functionality, so that (hopefully!) each patch will be individually mergeable.

Patch-set: 1