Give default string values to getSafeAttributes
This was inspired by a user report when using substitutions in the
start-message field; when they used "{change}" as per the extant
documentation it gave them a Python object representation in the
output; not the info they were looking for.

getSafeAttributes() makes an object with fields that are expanded as
strings for objects in model.py.  However, in doing this it also
looses any nice string representation of the represented object; so in
cases like this where the result is used as a string you don't get
anything useful.

Add a default argument to the Attributes() class that is the repr
string that should be returned for the wrapped attributes.  For
maximum usefulness, this expands the repr string with the incoming
arguments so you can use them to construct something helpful.
However, for most existing cases we can just pass the __repr__ string
of the original object here -- the fields represented are the "safe"
fields anyway, and some of the repr functions do non-trivial things to
make a nice string.

While this fixes the original issue in one way, we expand the
documentation on the start-msg to show the fields available and give
an example.  We add a test-case to cover this default expansion.

Change-Id: I741cd397a863512f06d941fd57c662aae55362e4