Add feature to exclude a branch pattern from tenant
When updating subtrees in gerrit is useful to create a branch containing
the squashed history of the upstream repositroy and then merge the subtree
change.
This workflow however causes issues with zuul if both projects use the same
path for storing zuul configuration, and will lead to multiple definitoins of
configuration shadowing eachother which leads to an error.

To allow for this workflow this change extends the tenant configuration to
tell zuul to skip loading any branches matching a regular expression.

Change-Id: I995311fa00accdbd2bd9a8b1be92a8ff38e26766