Update patch set 5
Patch Set 5: Code-Review-1

First a general comment: this seems like a good improvement -- but I had the idea that we wanted to eventually remove this functionality from this client and expect folks to use zuul-client instead.  If that's our goal, I don't think we should support this.  If it's not our goal then... erm... I'm sad that we have duplicate code and we will have a permanently confusing situation for users ("zuul client" vs "zuul-client" is already becoming difficult).

If we want behavior like this (so that admins don't have to generate tokens and make config files), but still also want to deprecate/remove these ops from the "zuul" command, then maybe we can add features to zuul-client so that if you run it on a zuul server (zuul.conf is available) it would do this?

Second: this is a real -1: if we do decide to do this, this should have docs updates -- probably the best thing would be to update the command's general help text to explain what it does.

Patch-set: 5
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1