Fix git error in stream jobs
New versions of git refuse to run on git repos owned by a user
other than the caller.  This is due to a fix for
https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-24765

The stream jobs install Zuul as root while the repo is owned by
zuul.  This causes pbr to run git which runs afoul of that issue.
To correct this, run a wheel build as the zuul user first and
install the wheel.

Change-Id: Id48245f715760c436ae46415f057358e1b687181