Fix background layout updates in zuul-web
Change Ia0a524053a110e8e48f709d219651e2ad9c8513d updated the scheduler
to more aggressively use the file cache, but did not make the corresponding
updates to zuul-web.  This means that configuration changes are not being
reflected in the web UI until the service restarts.

This corrects that and adds a test.

Change-Id: Iebdf485105bd02e57fa0d3db5ba162308b640ca0