Make nodepool hostvars available on unreachable hosts
Zuul ignores nodes with network_cli connections when running the
setup playbook in order to support the use-case where a network
appliance begins a job in an unreachable state.

Starting with Zuul v4.6.0, for security reasons, host variables
require a functioning Ansible connection in order to be set.
This includes the Nodepool variables such as public and private
IP addresses.  Causing a network_cli host to become online after
the start of a job requires this information, and so this
functionality has not worked since 4.6.0.

To correct this, we now always add in the nodepool host vars as
originally set regardless of whether the variable freeze playbook
runs for a particular host.  This means that even if we have no
other variables set for that host, we at least know the IP address
of the host and can interact with it directly in order to bring it
online.

Additionally, the freeze playbook did not correctly check the
return value, due to a typo.  This accidentally allowed Zuul to
continue to function in this situation (but without the nodepool
variables accessible).  With this change we continue to ignore
the return value, but intentionally so.

A test for this use-case is added, along with a releaso note.

Change-Id: Icd8a2c035e6c04a7c198281adbd07fef422a6c63
Story: 2009226