ansible zuul_json callback : handle string results
We have noted a problem with tasks such as "yum:" or "package:" that
output is not being logged in the job json file.

After investigation, I have found that this is failing with an
exception in the results parsing at

   for index, item_result in enumerate(
           clean_result.get('results', [])):
       if not item_result.get('_ansible_no_log', False):
           continue

When we are *not* processing a task loop, "results" is a simple list
of strings about the package results; thus when it walks each result
the "item_result.get('_ansible_no_log', False)" step fails because a
string has no ".get()".  This causes the plugin to exit and the
information in the resulting .json to be incomplete, leading to the
tasks being missing from the UI.  I believe this was a regression
introduced with I9e8d08f75207b362ca23457c44cc2f38ff43ac23.

When we *are* processing a task loop, "results" is a list where each
entry is a loop-item result.  In this case, we are walking a list of
dictionaries, so the ".get()" call works ... however ...

I could not not make the extant code fall into this path despite
trying most all combinations of loops and no_log flags.  Something
like

  - task:
    stat:
      path: '{{ item }}'
    loop:
      - /dev/null
      - /dev/zero
    no_log: true

Results in the callback getting "result._task.no_log == True" and
"result._result" being an already censored json dict.  i.e. I am not
convinced in the following

   if result._result.get('_ansible_no_log', False) or result._task.no_log:

that we will see an "_ansible_no_log" attribute in the result object
passed in.  Similarly we are checking the results of each loop
iteration for this attribute, which again I could not trigger with
current Ansible.

There are several related Ansible changes from around this time [1,2].
I've run out of energy to reverse engineer what's going on; I tried
rolling back to a 2.5-era Ansible to compare, but it wouldn't work
with CentOS 9 Python.  This interface is barely documented and Ansible
generally seems allergic to writing any sort of commit message that
gives context to changes.  I think we can leave it, out of an abudance
of caution, to avoid re-introducing a leak.  But I've left a note that
will lead back here, and we can use this as a starting point if we
have to dig further.

[1] https://github.com/ansible/ansible/commit/bda074d34e46ee9862a48ed067ad42260d3f92ab
[2] https://github.com/ansible/ansible/commit/336b3762b23a64e355cfa3efba11ddf5bdd7f0d8

Change-Id: I02bcd307bcfad8d99dd0db13d979ce7ba3d5e0e4