Further no_job reporting fixes
The previous fix for erroneously reporting NO_JOB results to the
database did not handle the case of superceding queues.  An item
enqueued in check and then superceded into gate would still report
NO_JOBS in gate and DEQUEUED in check, neither of which is desired.

It appears that we were previously relying on not having reported
a buildset start in order to avoid reporting buildset ends when
they are uninteresting.  This changed in
bf2eb71f95257e0dfac259fb74e7a97fe4a53eb8 where we now intentionally
create missing buildset entries.

To return to the original behavior, we now only attempt to report
a buildset end under most circumstances if we would have reported
a start as well (by virtue of having a non-empty job graph).  There
is one exception to this, where we report an item which otherwise
can't be enqueued in order to report an error.  This is maintained.

Change-Id: Ic2322e293a44a2946c6b766cf87d256ed39319ea