Update patch set 42
Patch Set 42: Code-Review-1

(1 comment)

-1 for inline comment.

Here are the pros and cons I see for Guillaume's change:

Pro: It works with all current and future drivers.
Pro: It is independent of the drivers themselves.
Con: It is complex (several git operations depending on various cases).
Con: It sometimes requires additional remote fetching.

And this change:

Pro: It is simple and easy to follow the logic.
Pro: It requires minimal changes to the merger.
Con: It requires per-driver support (but Gerrit, GitHub, and Gitlab are covered).
Con: It does not appear that Pagure has the required info.

Regarding the driver support: I have a followup change that will implement Gitlab support.  Gerrit doesn't need this because we use a new rest api to get the file list.  The only driver left out is Pagure, so the edge cases that Dong listed will apply to Pagure (at least unless and until they add a corresponding field to their api).

My inclination is to lean toward this solution since it is significantly simpler and relies on data from the remote system ot the greatest degree possible.

Patch-set: 42
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1