Update patch set 42
Patch Set 42:

I was mistaken about the Pagure driver, and we actually don't
need any changes there.  We get the files list from Pagure in all
cases by hitting the diffstat api endpoint, and it returns the
list we would expect (reverted files are not included), and I
verified that it returns at least 5000 entries (I haven't looked
at the source code to see if there's a limit, but that's a pretty
long tail of edge cases we're covering at this point).

I'm going to go ahead and push up a new patchset that addresses
my -1 and removes the print statements.

Here is a revised pro/con list:

Here are the pros and cons I see for Guillaume's change:                        
                                                                                
Pro: It works with all current and future drivers.                              
Pro: It is independent of the drivers themselves.                               
Con: It is complex (several git operations depending on various cases).         
Con: It sometimes requires additional remote fetching.                          
                                                                                
And this change:                                                                
                                                                                
Pro: It is simple and easy to follow the logic.                                 
Pro: It requires minimal changes to the merger.                                 
Con: It requires per-driver support (but all current drivers are covered).
                                                                                
My inclination is to lean toward this solution since it is
significantly simpler and relies on data from the remote system
ot the greatest degree possible.

Patch-set: 42